Strip the version from the Cargo.toml.
The version of rddns is determined by Git tags. The Cargo.toml as a second source for a version number only leaves room for divergence. Also setting a version in the Cargo.toml forces to decide for a version at the beginning of a development cycle. Its easier to decide if a release should be a new major or minor version at the end of such cycle when planning to actually release it.